33, flask assets

+++ app/routes/post.py
@@ -1,8 +1,8 @@
-from flask import Blueprint, render_template, request, redirect
+from flask import Blueprint, render_template, request, redirect, abort
 from flask_login import login_required, current_user
 
 from ..models.user import User
-from ..forms import StudentForm
+from ..forms import StudentForm, TeacherForm
 from ..extensions import db
 from ..models.post import Post
 
@@ -11,8 +11,17 @@ post = Blueprint('post', __name__)
 
 @post.route('/', methods=['GET', 'POST'])
 def all():
-    posts = Post.query.order_by(Post.date.desc()).all()
-    return render_template('post/all.html', posts=posts, user=User)
+    form = TeacherForm()
+    form.teacher.choices = [t.name for t in User.query.filter_by(status='teacher')]
+
+    if request.method == 'POST':
+        teacher = request.form.get('teacher')
+        teacher_id = User.query.filter_by(name=teacher).first().id
+        posts = Post.query.filter_by(teacher=teacher_id).order_by(Post.date.desc()).all()
+    else:
+        posts = Post.query.order_by(Post.date.desc()).all()
+
+    return render_template('post/all.html', posts=posts, user=User, form=form)
 
 
 @post.route('/post/create', methods=['POST', 'GET'])
@@ -40,9 +49,18 @@ def create():
 @login_required
 def update(id):
     post = Post.query.get_or_404(id)
+
+    if post.author != current_user:
+        abort(403)
+
+    form = StudentForm()
+    form.student.data = User.query.filter_by(id=post.student).first().name
+    form.student.choices = [s.name for s in User.query.filter_by(status='user')]
     if request.method == 'POST':
         post.subject = request.form.get('subject')
-        post.student = request.form.get('student')
+        student = request.form.get('student')
+
+        post.student = User.query.filter_by(name=student).first().id
         try:
             db.session.add(post)
             db.session.commit()
@@ -50,13 +68,17 @@ def update(id):
         except Exception as e:
             print(str(e))
     else:
-        return render_template('post/update.html', post=post)
+        return render_template('post/update.html', post=post, form=form)
 
 
 @post.route('/post/<int:id>/delete', methods=['POST', 'GET'])
 @login_required
 def delete(id):
     post = Post.query.get_or_404(id)
+
+    if post.author != current_user:
+        abort(403)
+
     try:
         db.session.delete(post)
         db.session.commit()
